requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
